docs(introduction): Add a narrative quick start
I wrote this as I worked, with the help of more knowledgeable folks on Discord, to get
up and running with the use case of feeding all announcements to Servarr.  I was
confused about what the larger overview of autobrr is, how it's parts are intended to
work together, and I couldn't find anything like that in the docs.  So I captured the
bits I was confused about as I encountered them and wrote them up in narrative form here
in the hopes it might help others with similar confusions.