docs(introduction): Network initially unhealthy
I just added another indexer with a different IRC network and noticed that autobrr
displays and logs the network unhealthy before anything has had a chance to go wrong.
This was very confusing for me so personally I wish that behavior was different but this
at least should help similar users understand better.